# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Service Management API",
        "description": "API documentation for Car Service / Car Wash Management System",
        "contact": {
            "name": "API Support",
            "email": "admin@carservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/vehicle-models": {
            "get": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "List vehicle models",
                "description": "Get paginated list of vehicle models",
                "operationId": "vehicleModelsIndex",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "vehicle_brand_id",
                        "in": "query",
                        "description": "Filter by vehicle brand",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle models retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "Create vehicle model",
                "description": "Create a new vehicle model",
                "operationId": "vehicleModelsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_brand_id",
                                    "name"
                                ],
                                "properties": {
                                    "vehicle_brand_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle model created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vehicle-models/by-brand/{vehicleBrandId}": {
            "get": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "List models by vehicle brand",
                "description": "Get all vehicle models for a specific vehicle brand",
                "operationId": "vehicleModelsListByBrand",
                "parameters": [
                    {
                        "name": "vehicleBrandId",
                        "in": "path",
                        "description": "Vehicle Brand ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle models retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vehicle-models/{id}": {
            "get": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "Get vehicle model",
                "description": "Get vehicle model details by ID",
                "operationId": "vehicleModelsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle Model ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle model retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle model not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "Update vehicle model",
                "description": "Update vehicle model details",
                "operationId": "vehicleModelsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle Model ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_brand_id",
                                    "name"
                                ],
                                "properties": {
                                    "vehicle_brand_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle model updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle model not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicle Models"
                ],
                "summary": "Delete vehicle model",
                "description": "Soft delete a vehicle model",
                "operationId": "vehicleModelsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle Model ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle model deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle model not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}